add tune target entropy scale

+++ refactor/algorithms/MAIC.py
@@ -50,11 +50,11 @@ class SAC_REINFORCE:
 		self.auto_entropy_tuning = auto_entropy_tuning
 		if self.auto_entropy_tuning:
 			self.target_entropy = float(self.num_agents * action_dim) * self.target_entropy_scale
-			self.log_alpha1 = torch.tensor([np.log(alpha1)], requires_grad=True, device=self.device)
+			self.log_alpha1 = torch.tensor([np.log(alpha1)], dtype=torch.float32, requires_grad=True, device=self.device)
 			self.alpha1_optimizer = optim.Adam([self.log_alpha1], lr=lr)
 			self.alpha1 = self.log_alpha1.exp()
 
-			self.log_alpha2 = torch.tensor([np.log(alpha2)], requires_grad=True, device=self.device)
+			self.log_alpha2 = torch.tensor([np.log(alpha2)], dtype=torch.float32, requires_grad=True, device=self.device)
 			self.alpha2_optimizer = optim.Adam([self.log_alpha2], lr=lr)
 			self.alpha2 = self.log_alpha2.exp()
 		else:
